Test for LocalProject.checkout_pr for GitLab
